_07_arrayBasedList/attempts/StackApplications.java
package practicePackage._07_arrayBasedList.attempts;

public class StackApplications {
	/**
	 * first read up on "reverse polish notation"
	 * https://en.wikipedia.org/wiki/Reverse_Polish_notation
	 * https://mathworld.wolfram.com/ReversePolishNotation.html
	 * http://www-stone.ch.cam.ac.uk/documentation/rrf/rpn.html
	 * @param expression
	 * assuming the passed String contains items in an reverse polish notation format,
	 * return it's evaluated value. Return null if the expression is invalid.
	 * For the basic version, you may assume each token is separated by a single space 
	 * and the expression is valid
	 * For the advanced version, there might be extra leading, trailing and intermediate spaces, 
	 * and the expression may or may not be valid. 
	 * 
	 * The operators possible are *, /, + and -
	 */
	public static Double evaluateRPN(String expression) {
		int length = expression.length();
		char[] arr = new char[length];
		//populate array of character
		for (int i = 0;i<length;i++) {
			arr[i] = expression.charAt(i);
			for(int k = 0
			//check if the 
		}
		return null;
	}
}
